pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "image-viewer"
version = "0.1.0"
description = "A lightweight image viewer with thumbnails, slideshow, ratings, and viewed tracking"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "Pillow>=9.0",
    "PyGObject>=3.42",
]

[project.scripts]
image-viewer = "image_viewer.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
